refactor: 요구사항에 맞게 package 수정
