+++ src/main/kotlin/racingcar/domain/RacingCarGenerator.kt
@@ -1,0 +1,12 @@
+package racingcar.domain
+
+import racingcar.domain.car.Car
+import racingcar.domain.car.CarName
+
+interface RacingCarGenerator {
+    fun generate(name: CarName): Car
+
+    companion object {
+        val STANDARD = StandardRacingCarGenerator()
+    }
+}
