+++ src/main/kotlin/racingcar/domain/car/CarLocation.kt
@@ -1,0 +1,17 @@
+package racingcar.domain.car
+
+import racingcar.domain.car.engine.Torque
+
+data class CarLocation(val value: Int) : Comparable<CarLocation> {
+    fun moveBy(torque: Torque): CarLocation {
+        return CarLocation(this.value + torque.value)
+    }
+
+    override fun compareTo(other: CarLocation): Int {
+        return this.value.compareTo(other.value)
+    }
+
+    companion object {
+        val ZERO = CarLocation(0)
+    }
+}
