+++ src/main/kotlin/racingcar/domain/car/CarName.kt
@@ -1,0 +1,11 @@
+package racingcar.domain.car
+
+data class CarName(val name: String) {
+    init {
+        require(name.length <= 5) {
+            "The length of name must be less than 6. name='$name'"
+        }
+    }
+
+    override fun toString() = name
+}
