+++ src/main/kotlin/racingcar/domain/car/engine/Engine.kt
@@ -1,0 +1,5 @@
+package racingcar.domain.car.engine
+
+interface Engine {
+    fun run(): Torque
+}
