+++ src/main/kotlin/racingcar/domain/car/engine/ForceGenerator.kt
@@ -1,0 +1,5 @@
+package racingcar.domain.car.engine
+
+interface ForceGenerator {
+    fun generate(): Force
+}
