+++ src/main/kotlin/racingcar/domain/car/engine/RandomForceGenerator.kt
@@ -1,0 +1,7 @@
+package racingcar.domain.car.engine
+
+class RandomForceGenerator : ForceGenerator {
+    override fun generate(): Force {
+        return Force((0..9).random())
+    }
+}
