+++ src/main/kotlin/racingcar/domain/car/engine/Torque.kt
@@ -1,0 +1,12 @@
+package racingcar.domain.car.engine
+
+/**
+ * [Engine]이 발생시키는 돌림힘.
+ *
+ * 차를 움직여 차의 위치를 변경시킨다.
+ */
+enum class Torque(val value: Int) {
+    STOP(0),
+    MOVE(1)
+    ;
+}
