--- src/main/kotlin/racingcar/racing/RacingCarGenerator.kt
@@ -1,12 +1,0 @@
-package racingcar.racing
-
-import racingcar.racing.car.Car
-import racingcar.racing.car.CarName
-
-interface RacingCarGenerator {
-    fun generate(name: CarName): Car
-
-    companion object {
-        val STANDARD = StandardRacingCarGenerator()
-    }
-}
