--- src/main/kotlin/racingcar/racing/RacingResults.kt
@@ -1,15 +1,0 @@
-package racingcar.racing
-
-class RacingResults(trials: List<RacingTrial>) {
-    val trials: List<RacingTrial> = trials.sortedBy { it.order }
-
-    init {
-        this.trials.forEachIndexed { index, trial -> trial.checkOrder(index) }
-    }
-
-    val winnerRecords: List<Record>
-        get() = lastTrial.leadRecords
-
-    private val lastTrial: RacingTrial
-        get() = trials.last()
-}
