--- src/main/kotlin/racingcar/racing/car/Car.kt
@@ -1,19 +1,0 @@
-package racingcar.racing.car
-
-import racingcar.racing.car.engine.Engine
-
-class Car(
-    val name: CarName,
-    private val engine: Engine
-) {
-    var location: CarLocation = CarLocation.ZERO
-
-    fun race(): CarLocation {
-        goForward()
-        return location
-    }
-
-    private fun goForward() {
-        location = location.moveBy(engine.run())
-    }
-}
