--- src/main/kotlin/racingcar/racing/car/CarName.kt
@@ -1,11 +1,0 @@
-package racingcar.racing.car
-
-data class CarName(val name: String) {
-    init {
-        require(name.length <= 5) {
-            "The length of name must be less than 6. name='$name'"
-        }
-    }
-
-    override fun toString() = name
-}
