--- src/main/kotlin/racingcar/racing/car/engine/Engine.kt
@@ -1,5 +1,0 @@
-package racingcar.racing.car.engine
-
-interface Engine {
-    fun run(): Torque
-}
