--- src/main/kotlin/racingcar/racing/car/engine/Force.kt
@@ -1,15 +1,0 @@
-package racingcar.racing.car.engine
-
-data class Force(val value: Int) {
-    init {
-        require(value >= ZERO_VALUE) { "Force must be positive or zero." }
-    }
-
-    operator fun compareTo(other: Force): Int {
-        return this.value.compareTo(other.value)
-    }
-
-    companion object {
-        private const val ZERO_VALUE = 0
-    }
-}
