--- src/main/kotlin/racingcar/racing/car/engine/ForceGenerator.kt
@@ -1,5 +1,0 @@
-package racingcar.racing.car.engine
-
-interface ForceGenerator {
-    fun generate(): Force
-}
