--- src/main/kotlin/racingcar/racing/car/engine/RandomForceGenerator.kt
@@ -1,7 +1,0 @@
-package racingcar.racing.car.engine
-
-class RandomForceGenerator : ForceGenerator {
-    override fun generate(): Force {
-        return Force((0..9).random())
-    }
-}
